backend/requirements.txt
fastapi
uvicorn[standard]
pandas
python-multipart
fpdf2